Add core module
