+++ packages/backend-resolver/src/server_builder.ts
@@ -5,7 +5,7 @@ import {getShortLink} from "@xivgear/gearplan-frontend/external/shortlink_server
 import {PartyBonusAmount, SheetStatsExport} from "@xivgear/xivmath/geartypes";
 import {getBisSheet} from "@xivgear/gearplan-frontend/external/static_bis";
 import {registerDefaultSims} from "@xivgear/gearplan-frontend/sims/default_sims";
-import {GearPlanSheet} from "@xivgear/gearplan-frontend/sheet";
+import {HEADLESS_SHEET_PROVIDER} from "@xivgear/gearplan-frontend/sheet";
 
 let initDone = false;
 function checkInit() {
@@ -22,7 +22,7 @@ function doInit() {
 async function importExportSheet(request, rawData: string): Promise<SheetStatsExport> {
     const exported = JSON.parse(rawData);
     const isFullSheet = 'sets' in exported;
-    const sheet = isFullSheet ? GearPlanSheet.fromExport(exported) : GearPlanSheet.fromSetExport(exported);
+    const sheet = isFullSheet ? HEADLESS_SHEET_PROVIDER.fromExport(exported) : HEADLESS_SHEET_PROVIDER.fromSetExport(exported);
     sheet.setViewOnly();
     const pb = request.query['partyBonus'];
     if (pb) {
