+++ packages/core/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@xivgear/core",
+  "version": "1.0.0",
+  "description": "",
+  "scripts": {
+    "build": "tsc --build",
+    "test": "ts-mocha src/test/**/*test.ts src/test/**/*tests.ts"
+  },
+  "devDependencies": {
+    "ts-node": "^10.9.2",
+    "ts-loader": "^9.4.4",
+    "ts-mocha": "^10.0.0",
+    "chai": "^4.4.1",
+    "@types/chai": "^4.3.16"
+  },
+  "exports": {
+    "./*": "./src/*.ts"
+  },
+  "files": [
+    "package.json",
+    "tsconfig.json",
+    "src/*.ts"
+  ]
+}
