+++ packages/core/src/test/fail_on_purpose_test.ts
@@ -1,0 +1,7 @@
+import {assert} from "chai";
+
+// describe('stuff', () => {
+//     it('fails on purpose to make sure the test is running', () => {
+//         assert.fail('bad')
+//     })
+// })
