+++ packages/frontend/package.json
@@ -3,7 +3,8 @@
   "version": "1.0.0",
   "dependencies": {
     "@webcomponents/webcomponentsjs": "^2.8.0",
-    "@xivgear/xivmath": "^1.0.0"
+    "@xivgear/xivmath": "^1.0.0",
+    "@xivgear/core": "^1.0.0"
   },
   "exports": {
     "./*": "./src/scripts/*.ts"
