+++ packages/frontend/src/scripts/base_ui.ts
@@ -10,7 +10,7 @@ import {SheetPickerTable} from "./components/saved_sheet_picker";
 import {DISPLAY_SETTINGS} from "./settings/display_settings";
 import {showSettingsModal} from "./settings/settings_modal";
 import {GearPlanSheet} from "./sheet";
-import {GearPlanSheetGui} from "./components/sheet";
+import {GearPlanSheetGui, GRAPHICAL_SHEET_PROVIDER} from "./components/sheet";
 
 const pageTitle = 'XivGear - FFXIV Gear Planner';
 
@@ -104,7 +104,7 @@ export async function openSheetByKey(sheet: string) {
     // TODO: handle nonexistent sheet
     setTitle('Loading Sheet');
     console.log('openSheetByKey: ', sheet);
-    const planner = GearPlanSheetGui.fromSaved(sheet);
+    const planner = GRAPHICAL_SHEET_PROVIDER.fromSaved(sheet);
     if (planner) {
         await openSheet(planner);
     }
@@ -116,7 +116,7 @@ export async function openSheetByKey(sheet: string) {
 
 export async function openExport(exported: (SheetExport | SetExport), changeHash: boolean, viewOnly: boolean) {
     const isFullSheet = 'sets' in exported;
-    const sheet = isFullSheet ? GearPlanSheetGui.fromExport(exported) : GearPlanSheetGui.fromSetExport(exported);
+    const sheet = isFullSheet ? GRAPHICAL_SHEET_PROVIDER.fromExport(exported) : GRAPHICAL_SHEET_PROVIDER.fromSetExport(exported);
     if (isEmbed() && !isFullSheet) {
         sheet.setViewOnly();
         openEmbed(sheet);
