+++ packages/frontend/src/scripts/components/saved_sheet_picker.ts
@@ -130,3 +130,4 @@ function startNewSheet() {
     showNewSheetForm();
 }
 
+customElements.define("gear-sheet-picker", SheetPickerTable, {extends: "table"});
