+++ packages/frontend/src/scripts/simulation.ts
@@ -1,8 +1,6 @@
 import {CharacterGearSet} from "./gear";
 import {JobName, SupportedLevel} from "@xivgear/xivmath/xivconstants";
 import {potRatioSimSpec} from "./sims/common/potency_ratio";
-import {CustomTable} from "./tables";
-import {camel2title} from "./util/strutils";
 
 /**
  * Represents the final result of a simulation run. Sim implementors should extend this type with
@@ -187,81 +185,6 @@ export interface Simulation<ResultType extends SimResult, SettingsType extends S
      * may also be implemented for changes to gear sets.
      */
     readonly manualRun?: boolean;
-}
-
-type SimpleResultEntry = {
-    name: string;
-    value: any;
-}
-
-export function bestEffortFormat(value: any): Node {
-    if (typeof value === 'number') {
-        return document.createTextNode(value.toFixed(3));
-    }
-    else {
-        return document.createTextNode(value.toString());
-    }
-}
-
-/**
- * Simple table for displaying key/value pairs of an object. The left column is the
- * keys, while the right column is the values. No header row.
- *
- * @param result The result to display
- */
-export function simpleAutoResultTable(result: Object): HTMLElement {
-    const data: SimpleResultEntry[] = [];
-    for (let fieldKey in result) {
-        data.push({
-            name: camel2title(fieldKey),
-            value: result[fieldKey]
-        });
-    }
-    const table = new CustomTable<SimpleResultEntry>();
-    table.columns = [
-        {
-            shortName: 'key',
-            displayName: 'Key',
-            getter: item => item.name,
-        },
-        {
-            shortName: 'value',
-            displayName: 'Value',
-            getter: item => item.value,
-            renderer: bestEffortFormat,
-        }
-    ]
-    table.data = data;
-    return table;
-
-}
-
-export function simpleMappedResultTable<X extends SimResult>(fieldNames: { [K in keyof X]: string }): ((result: X) => HTMLElement) {
-    return (result: X): HTMLElement => {
-        const data: SimpleResultEntry[] = [];
-        for (let fieldKey in fieldNames) {
-            data.push({
-                name: fieldNames[fieldKey],
-                value: result[fieldKey]
-            });
-        }
-        const table = new CustomTable<SimpleResultEntry>();
-        table.columns = [
-            {
-                shortName: 'key',
-                displayName: 'Key',
-                getter: item => item.name,
-            },
-            {
-                shortName: 'value',
-                displayName: 'Value',
-                getter: item => item.value,
-                renderer: bestEffortFormat,
-            }
-        ]
-        table.data = data;
-        return table;
-    }
 }
 
 /**
